requirements.txt
fastapi[standard]
pydantic_settings
tortoise-orm[asyncpg]
firebase-admin
pyrebase4
python-jose[cryptography]
setuptools
reportlab